% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_info.R
\name{company_info}
\alias{company_info}
\title{Company Information}
\usage{
company_info()
}
\value{
A list of useful information about SpaceX
\describe {
  \item{name}{Company name}
  \item{founder}{Founder's name}
  \item{founded}{Year founded}
  \item{employees}{Number of employees}
  \item{vehicles}{Number of main types of launch vehicles}
  \item{launch_sites}{Number of launch sites}
  \item{ceo}{Name of the current CEO}
  \item{cto}{Name of the current CTO}
  \item{coo}{Name of the current COO}
  \item{cto_propulsion}{Name of the current CTO for Propulsion}
  \item{valuation}{Current company valuation (in USD)}
  \item{headquarters}{List of address components for the company headquarters}
  \item{summary}{Brief company description}
}
}
\description{
Provides general SpaceX company information
}
\examples{
company_info()

}
