% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicle_info.R
\name{vehicle_info}
\alias{vehicle_info}
\title{Vehicle Information}
\usage{
vehicle_info(id = NULL)
}
\arguments{
\item{id}{If not provided, gives a list of all vehicles. Otherwise,
just the specified vehicle.}
}
\value{
A list vehicle parameters
}
\description{
Provides information and stats on SpaceX company vehicles, both launch
vehicles and capsules
}
\examples{
vehicle_info("falcon9")

}
